+++ RELEASE-NOTES.md
@@ -1,5 +1,9 @@
 ## 1.5.1
 
+### 新功能
+
+1. [ISSUE #314](https://github.com/dangdangdotcom/sharding-jdbc/issues/314) 支持DDL类型的SQL
+
 ### 缺陷修正
 
 1. [ISSUE #308](https://github.com/dangdangdotcom/sharding-jdbc/issues/308) 数据库原生的自增GeneratedKey的返回无效