+++ RELEASE-NOTES.md
@@ -4,6 +4,10 @@
 
 1. [ISSUE #314](https://github.com/dangdangdotcom/sharding-jdbc/issues/314) 支持DDL类型的SQL
 
+### 功能调整
+
+1. [ISSUE #327](https://github.com/dangdangdotcom/sharding-jdbc/issues/327) 默认关闭sql.show配置
+
 ### 缺陷修正
 
 1. [ISSUE #308](https://github.com/dangdangdotcom/sharding-jdbc/issues/308) 数据库原生的自增GeneratedKey的返回无效